# docs/api/openapi-specification.yaml
openapi: 3.0.3
info:
  title: オンライン診療システム API
  description: |
    ファストドクター主催「オンライン診療AI活用ハッカソン」のベースシステムAPI仕様書です。
    患者と医療従事者（医師・オペレータ・管理者）の両方に対応したRESTful APIを提供します。
  version: 1.0.0
  contact:
    name: ハッカソンサポートチーム
    email: hackathon@fastdoctor.jp
servers:
  - url: http://localhost:8787
    description: ローカル開発環境
  - url: https://api.online-medical.example.com
    description: 本番環境（Cloudflare Workers）
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT認証トークン
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: エラーコード
        message:
          type: string
          description: エラーメッセージ
        details:
          type: object
          description: 詳細なエラー情報
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 総件数
        page:
          type: integer
          description: 現在のページ番号
        limit:
          type: integer
          description: 1ページあたりの件数
        hasNext:
          type: boolean
          description: 次のページが存在するか
        hasPrev:
          type: boolean
          description: 前のページが存在するか

    # ユーザー関連
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        address:
          type: string
        emergencyContact:
          type: string
        medicalHistory:
          type: array
          items:
            type: object
            properties:
              condition:
                type: string
              diagnosedDate:
                type: string
                format: date
              notes:
                type: string
        profileImageUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Medication:
      type: object
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        instructions:
          type: string
    Prescription:
      type: object
      properties:
        id:
          type: integer
        appointmentId:
          type: integer
        workerId:
          type: integer
        medications:
          type: array
          items:
            $ref: '#/components/schemas/Medication'
        pharmacyNotes:
          type: string
        status:
          type: string
          enum: [draft, issued, dispensed, cancelled]
        issuedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # チャット関連
...
  responses:
...
    ForbiddenError:
      description: 権限エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: FORBIDDEN
            message: このリソースへのアクセス権限がありません
    NotFoundError:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NOT_FOUND
            message: 指定されたリソースが見つかりません
...
security:
  - BearerAuth: []
paths:
...
  /api/patient/prescriptions:
    get:
      tags:
        - Patient
      summary: 処方箋一覧
      description: 自分の処方箋一覧を取得します
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [issued, dispensed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Prescription'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /api/worker/patients:
    get:
      tags:
        - Worker
      summary: 患者一覧取得
      description: 患者一覧を取得します
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: 名前・メールアドレス・電話番号で検索
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Patient'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/worker/patients/{id}:
    get:
      tags:
        - Worker
      summary: 患者詳細取得
      description: 指定した患者の詳細情報を取得します
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Worker
      summary: 患者情報更新
      description: 患者情報を更新します（管理者のみ）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phoneNumber:
                  type: string
                address:
                  type: string
                emergencyContact:
                  type: string
                medicalHistory:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...